feat(chatbot): messagestore use immer
